add validations

+++ src/routes.js
@@ -1,6 +1,7 @@
 import { Database } from "./database.js"
 import { buildRoutePath } from "./utils/build-route-path.js"
 import { randomUUID } from "crypto"
+import { isTaskEmpty, taskFound, taskCompleted } from "./utils/task-validator.js"
 
 const database = new Database()
 
@@ -10,6 +11,11 @@ export const routes = [
     path: buildRoutePath("/tasks"),
     handler: (req, res) => {
       const { title, description } = req.body
+
+      if (!isTaskEmpty(req.body)) {
+        return res.writeHead(400).end("Task title and description are required")
+      }
+
       const task = {
         id: randomUUID(),
         title,
@@ -39,6 +45,11 @@ export const routes = [
     path: buildRoutePath("/tasks/:id"),
     handler: (req, res) => {
       const { id } = req.params
+      
+      if(!taskFound(id)) {
+        return res.writeHead(404).end("Task not found")
+      }
+
       database.delete("tasks", id)
       return res.writeHead(204).end()
     },
@@ -49,6 +60,11 @@ export const routes = [
     handler: (req, res) => {
       const { id } = req.params
       const { title, description } = req.body
+      
+      if(!taskFound(id)) {
+        return res.writeHead(404).end("Task not found")
+      }
+
       const task = database.select("tasks", { id })[0]
       const updatedTask = {
         ...task,
@@ -65,6 +81,15 @@ export const routes = [
     path: buildRoutePath("/tasks/:id/complete"),
     handler: (req, res) => {
       const { id } = req.params
+
+      if(!taskFound(id)) {
+        return res.writeHead(404).end("Task not found")
+      }
+
+      if(taskCompleted(id)) {
+        return res.writeHead(400).end("Task already completed")
+      }
+
       const task = database.select("tasks", { id })[0]
       const updatedTask = {
         ...task,
